+++ PoGo.NecroBot.CLI/BotDataSocketClient.cs
@@ -180,12 +180,13 @@ namespace PoGo.NecroBot.CLI
             {
                 OnSnipePokemon(session, e.Data);
                 OnPokemonData(session, e.Data);
-                ONFPMBridgeData(session, e.Data);
+                //ONFPMBridgeData(session, e.Data); //Nolonger use
             }
+
             catch (Exception ex)
             {
 #if DEBUG
-                Logger.Write("ERROR TO ADD SNIPE< DEBUG ONLY " + ex.Message, LogLevel.Info, ConsoleColor.Yellow);
+                Logger.Write("ERROR TO ADD SNIPE< DEBUG ONLY " + ex.Message + "\r\n " + ex.StackTrace, LogLevel.Info, ConsoleColor.Yellow);
 #endif
             }
 
@@ -201,10 +202,11 @@ namespace PoGo.NecroBot.CLI
             }
         }
 
-        public static bool CheckIfPokemonBeenCaught(double lat, double lng, PokemonId id , ISession session)
+        public static bool CheckIfPokemonBeenCaught(double lat, double lng, PokemonId id , ulong encounterId, ISession session)
         {
             string uniqueCacheKey = $"{session.Settings.PtcUsername}{session.Settings.GoogleUsername}{Math.Round(lat, 6)}{id}{Math.Round(lng, 6)}";
             if (session.Cache.Get(uniqueCacheKey) != null) return true;
+            if (encounterId > 0 && session.Cache[encounterId.ToString()] !=null) return true;
 
             return false;
 
@@ -219,13 +221,13 @@ namespace PoGo.NecroBot.CLI
                 session.EventDispatcher.Send(data);
                 if (session.LogicSettings.AllowAutoSnipe)
                 {
-                    var move1 = PokemonMove.Absorb;
-                    var move2 = PokemonMove.Absorb;
+                    var move1 = PokemonMove.MoveUnset;
+                    var move2 = PokemonMove.MoveUnset;
                     Enum.TryParse<PokemonMove>(data.Move1, true, out move1);
                     Enum.TryParse<PokemonMove>(data.Move1, true, out move2);
                     ulong encounterid = 0;
                     ulong.TryParse(data.EncounterId, out encounterid);
-                    bool caught = CheckIfPokemonBeenCaught(data.Latitude, data.Longitude, data.PokemonId, session);
+                    bool caught = CheckIfPokemonBeenCaught(data.Latitude, data.Longitude, data.PokemonId, encounterid, session);
                     if (!caught)
                     {
                         MSniperServiceTask.AddSnipeItem(session, new MSniperServiceTask.MSniperInfo2()
@@ -237,8 +239,9 @@ namespace PoGo.NecroBot.CLI
                             PokemonId = (short)data.PokemonId,
                             Iv = data.IV,
                             Move1 = move1,
-                            Move2 = move2
-                        });
+                            Move2 = move2       ,
+                            ExpiredTime =data.ExpireTimestamp
+                        }).Wait();
                     }
                 
                 }
@@ -260,8 +263,10 @@ namespace PoGo.NecroBot.CLI
                 var move2 = PokemonMove.Absorb;
                 Enum.TryParse<PokemonMove>(data.Move1, true, out move1);
                 Enum.TryParse<PokemonMove>(data.Move1, true, out move2);
+                ulong encounterid = 0;
+                ulong.TryParse(data.EncounterId, out encounterid);
 
-                bool caught = CheckIfPokemonBeenCaught(data.Latitude, data.Longitude, data.PokemonId, session);
+                bool caught = CheckIfPokemonBeenCaught(data.Latitude, data.Longitude, data.PokemonId, encounterid, session);
                 if(caught)
                 {
                     Logger.Write("[SNIPE IGNORED] - Your snipe pokemon has already been cautgh by bot", LogLevel.Sniper);
@@ -272,13 +277,14 @@ namespace PoGo.NecroBot.CLI
                 {
                     Latitude = data.Latitude,
                     Longitude = data.Longitude,
-                    EncounterId = data.EncounterId.Contains("-")?0:Convert.ToUInt64(data.EncounterId),
+                    EncounterId = encounterid,
                     SpawnPointId = data.SpawnPointId,
                     PokemonId = (short)data.PokemonId,
                     Iv = data.IV,
                     Move1 = move1,
+                    ExpiredTime = data.ExpireTimestamp        ,
                     Move2 = move2
-                }, true);
+                }, true).Wait();
             }
         }