+++ PoGo.NecroBot.CLI/ConsoleEventListener.cs
@@ -42,7 +42,6 @@ namespace PoGo.NecroBot.CLI
             if (!warnEvent.RequireInput) return;
             Logger.Write(session.Translation.GetTranslation(TranslationString.RequireInputText), LogLevel.Warning);
         }
-
         private static void HandleEvent(UseLuckyEggEvent useLuckyEggEvent, ISession session)
         {
             Logger.Write(session.Translation.GetTranslation(TranslationString.EventUsedLuckyEgg, useLuckyEggEvent.Count),
@@ -52,11 +51,12 @@ namespace PoGo.NecroBot.CLI
         private static void HandleEvent(PokemonEvolveEvent pokemonEvolveEvent, ISession session)
         {
             string strPokemon = session.Translation.GetPokemonTranslation(pokemonEvolveEvent.Id);
-            Logger.Write(pokemonEvolveEvent.Result == EvolvePokemonResponse.Types.Result.Success
+            string logMessage = pokemonEvolveEvent.Result == EvolvePokemonResponse.Types.Result.Success
                 ? session.Translation.GetTranslation(TranslationString.EventPokemonEvolvedSuccess, strPokemon, pokemonEvolveEvent.Exp)
                 : session.Translation.GetTranslation(TranslationString.EventPokemonEvolvedFailed, pokemonEvolveEvent.Id, pokemonEvolveEvent.Result,
-                    strPokemon),
-                LogLevel.Evolve);
+                    strPokemon);
+            logMessage = (pokemonEvolveEvent.Sequence > 0 ? $"{pokemonEvolveEvent.Sequence}. " : "") + logMessage; 
+            Logger.Write(logMessage, LogLevel.Evolve);
         }
 
         private static void HandleEvent(TransferPokemonEvent transferPokemonEvent, ISession session)