+++ PoGo.NecroBot.CLI/Properties/AssemblyInfo.cs
@@ -40,7 +40,7 @@ using System.Runtime.InteropServices;
 // [assembly: AssemblyVersion1("0.9.9.5")]
 
 
-[assembly: AssemblyVersion("1.0.0.37")]
+[assembly: AssemblyVersion("1.0.0.51")]
 [assembly: AssemblyFileVersion("1.0.0.0")]
-[assembly: AssemblyInformationalVersion("v1.0.0.37")]
+[assembly: AssemblyInformationalVersion("v1.0.0.51")]