+++ PoGo.NecroBot.Logic/Common/Translations.cs
@@ -270,7 +270,9 @@ namespace PoGo.NecroBot.Logic.Common
         FailedSendNotification,
         TelegramBotStarted,
         TelegramNeedChatId,
-        BulkTransferFailed
+        BulkTransferFailed,
+        AutoSnipeDisabled,
+        SnipePokemonNotInPokedex
     }
 
     public class Translation : ITranslation
@@ -494,6 +496,8 @@ namespace PoGo.NecroBot.Logic.Common
                 "You need to fill out PtcUsername and PtcPassword in auth.json!"),
             new KeyValuePair<TranslationString, string>(TranslationString.SnipeScan,
                 "Scanning for Snipeable Pokemon at {0}..."),
+            new KeyValuePair<TranslationString, string>(TranslationString.SnipePokemonNotInPokedex,
+                "Auto sniper detected a pokemon not in your pokedex:  {0}. He will be snipped as priority!"),
             new KeyValuePair<TranslationString, string>(TranslationString.SnipeScanEx,
                 "Sniping a {0} with {1} IV at {2}..."),
             new KeyValuePair<TranslationString, string>(TranslationString.NoPokemonToSnipe,
@@ -596,6 +600,7 @@ namespace PoGo.NecroBot.Logic.Common
             new KeyValuePair<TranslationString, string>(TranslationString.AccountBanned, "Probably Permanent Ban!"),
             new KeyValuePair<TranslationString, string>(TranslationString.GoogleAPIWarning, "Without a Google Api, you will have 2500 free quota limit, if you reach the maximum quota, try to change your IP. To configure \"GoogleAPIKey\", get API Key in link: https://developers.google.com/maps/documentation/directions/get-api-key"),
             new KeyValuePair<TranslationString, string>(TranslationString.Only10kmEggs, "Player below level 20, saving this 10 km Egg for later"),
+            new KeyValuePair<TranslationString, string>(TranslationString.AutoSnipeDisabled,"Your are out of ball because snipe so fast, you can reduce snipe speed by update MinIVForAutoSnipe or SnipePokemonFilters, Auto snipe will be disable in {0} mins"),
             new KeyValuePair<TranslationString, string>(TranslationString.SniperCount, "Sniper count {0}"),
             new KeyValuePair<TranslationString, string>(TranslationString.SnipeExceeds, "Sniper need to take a rest before your account is banned"),
             new KeyValuePair<TranslationString, string>(TranslationString.CatchExceeds, "You are catching too fast. Your cannot catch another one until {0} seconds later"),