+++ PoGo.NecroBot.Logic/Interfaces/Configuration/ILogicSettings.cs
@@ -143,7 +143,7 @@ namespace PoGo.NecroBot.Logic.Interfaces.Configuration
         bool VerboseRecycling { get; }
         double RecycleInventoryAtUsagePercentage { get; }
         double EvolveKeptPokemonsAtStorageUsagePercentage { get; }
-        int EvolveKeptPokemonsOverrideStartIfThisManyReady { get; }
+        int EvolveKeptPokemonIfBagHasOverThisManyPokemon { get; }
         bool UseSnipeLimit { get; }
         bool UsePokeStopLimit { get; }
         bool UseCatchLimit { get; }
@@ -240,6 +240,10 @@ namespace PoGo.NecroBot.Logic.Interfaces.Configuration
         List<AuthConfig> Bots { get; }
         bool AllowMultipleBot { get; }
         CaptchaConfig CaptchaConfig { get;  }
-        int BulkTransferStogareBuffer { get;}
+        int BulkTransferStogareBuffer { get;  }
+        int BulkTransferSize { get; }
+
+        bool AutosnipeVerifiedOnly { get;  }
+        int SnipePauseOnOutOfBallTime { get;  }
     }
 }