+++ PoGo.NecroBot.Logic/Inventory.cs
@@ -42,6 +42,13 @@ namespace PoGo.NecroBot.Logic
         {
             _client = client;
             _logicSettings = logicSettings;
+            //Inventory update will be call everytime getMabObject call
+            client.OnInventoryUpdated += (refreshedInventoryData) =>
+             {
+                 //Console.WriteLine("################# INVENTORY UPDATE ######################");
+                 _cachedInventory = refreshedInventoryData;
+                 _lastRefresh = DateTime.Now;
+             };
         }
 
         private Caching.LRUCache<ItemId, int> pokeballCache = new Caching.LRUCache<ItemId, int>(capacity: 10);
@@ -62,7 +69,7 @@ namespace PoGo.NecroBot.Logic
             ItemId.ItemMaxPotion
         };
 
-        public async Task UpdateInventoryItem(ItemId itemId, int count)
+        public void UpdateInventoryItem(ItemId itemId, int count)
         {
             foreach (var item in this._cachedInventory.InventoryDelta.InventoryItems)
             {
@@ -106,7 +113,7 @@ namespace PoGo.NecroBot.Logic
             if (_player == null) GetPlayerData();
             var now = DateTime.UtcNow;
 
-            if (_cachedInventory != null && _lastRefresh.AddSeconds(5*60).Ticks > now.Ticks)
+            if (_cachedInventory != null && _lastRefresh.AddSeconds(5 * 60).Ticks > now.Ticks)
                 return _cachedInventory;
 
             return await RefreshCachedInventory();
@@ -159,7 +166,7 @@ namespace PoGo.NecroBot.Logic
             {
                 throw e;
             }
-            catch (Exception e)
+            catch (Exception)
             {
                 //throw e; 
             }
@@ -249,11 +256,13 @@ namespace PoGo.NecroBot.Logic
 
         public async Task UpdateCandy(Candy family, int change)
         {
+            await Task.Delay(0); // Just added to get rid of compiler warning. Remove this if async code is used below.
+
             var lookupItem = (from item in _cachedInventory.InventoryDelta.InventoryItems
-                             where item.InventoryItemData?.Candy != null
-                             where item.InventoryItemData?.Candy.FamilyId == family.FamilyId
-                             select item.InventoryItemData.Candy).FirstOrDefault();
-            if(lookupItem == null && change >0)
+                              where item.InventoryItemData?.Candy != null
+                              where item.InventoryItemData?.Candy.FamilyId == family.FamilyId
+                              select item.InventoryItemData.Candy).FirstOrDefault();
+            if (lookupItem == null && change > 0)
             {
                 family.Candy_ = change;
                 _cachedInventory.InventoryDelta.InventoryItems.Add(new InventoryItem()
@@ -265,7 +274,7 @@ namespace PoGo.NecroBot.Logic
                 });
             }
             else
-            lookupItem.Candy_ += change;
+                lookupItem.Candy_ += change;
         }
 
         public async Task<IEnumerable<EggIncubator>> GetEggIncubators()
@@ -424,8 +433,8 @@ namespace PoGo.NecroBot.Logic
 
         public async Task<List<InventoryItem>> GetPokeDexItems()
         {
-            List<InventoryItem> PokeDex = new List<InventoryItem>();
-            var inventory = await _client.Inventory.GetInventory();
+            //List<InventoryItem> PokeDex = new List<InventoryItem>();
+            var inventory = await GetCachedInventory(); 
 
             return (from items in inventory.InventoryDelta.InventoryItems
                     where items.InventoryItemData?.PokedexEntry != null
@@ -474,8 +483,23 @@ namespace PoGo.NecroBot.Logic
                     PokemonData = data
                 }
             });
-        }
 
+            var pokedexEntry = (await GetPokeDexItems()).FirstOrDefault(x => x.InventoryItemData?.PokedexEntry?.PokemonId == data.PokemonId);
+
+            if (pokedexEntry == null)
+            {
+                _cachedInventory.InventoryDelta.InventoryItems.Add(new InventoryItem()
+                {
+                    InventoryItemData = new InventoryItemData()
+                    {
+                        PokedexEntry = new PokedexEntry()
+                        {
+                            PokemonId = data.PokemonId
+                        }
+                    }
+                });
+            }
+        }
         public async Task<PokemonData> GetSinglePokemon(ulong id)
         {
             var inventory = await GetCachedInventory();
@@ -504,12 +528,22 @@ namespace PoGo.NecroBot.Logic
             return
                 inventory.Where(i => !string.IsNullOrEmpty(i.DeployedFortId));
         }
+        public async Task<MoveSettings> GetMoveSetting(PokemonMove move)
+        {
+            if (_templates == null) await GetPokemonSettings();
 
+            var moveSettings = _templates.ItemTemplates.Where(x => x.MoveSettings != null).Select(x => x.MoveSettings).ToList();
+
+            return moveSettings.FirstOrDefault(x => x.MovementId == move);
+
+        }
         public async Task<IEnumerable<PokemonSettings>> GetPokemonSettings()
         {
             if (_templates == null || _pokemonSettings == null)
             {
                 _templates = await _client.Download.GetItemTemplates();
+                var moveSettings = _templates.ItemTemplates.Where(x => x.MoveSettings != null).Select(x => x.MoveSettings).ToList();
+
                 _pokemonSettings = _templates.ItemTemplates.Select(i => i.PokemonSettings).Where(p => p != null && p.FamilyId != PokemonFamilyId.FamilyUnset);
             }
             return _pokemonSettings;
@@ -572,11 +606,11 @@ namespace PoGo.NecroBot.Logic
             if (_level == 0 || level > _level)
             {
                 _level = level;
-                 
+
                 var rewards = await _client.Player.GetLevelUpRewards(level);
                 foreach (var item in rewards.ItemsAwarded)
                 {
-                    await UpdateInventoryItem(item.ItemId, item.ItemCount);
+                    UpdateInventoryItem(item.ItemId, item.ItemCount);
                 }
             }