+++ PoGo.NecroBot.Logic/Model/Settings/SnipeConfig.cs
@@ -120,8 +120,18 @@ namespace PoGo.NecroBot.Logic.Model.Settings
         public bool ActivateMSniper = true;
 
         [ExcelConfig(Description = "Min IV that bot will automatically snipe pokemon", Position = 21)]
-        [DefaultValue(95)]
+        [DefaultValue(100)]
         [JsonProperty(Required = Required.DisallowNull, DefaultValueHandling = DefaultValueHandling.Populate, Order = 21)]
         public int MinIVForAutoSnipe { get; set; }
+
+        [ExcelConfig(Description = "Only auto snipe pokemon has been verified (overwriteable by invidual pokemon)", Position = 22)]
+        [DefaultValue(false)]
+        [JsonProperty(Required = Required.DisallowNull, DefaultValueHandling = DefaultValueHandling.Populate, Order = 22)]
+        public bool AutosnipeVerifiedOnly { get; set; }
+
+        [ExcelConfig(Description = "Total time in minutes bot will ignore autosnipe when out of ball", Position = 23)]
+        [DefaultValue(5)]
+        [JsonProperty(Required = Required.DisallowNull, DefaultValueHandling = DefaultValueHandling.Populate, Order = 23)]
+        public int SnipePauseOnOutOfBallTime { get;  set; }
     }
 }