+++ PoGo.NecroBot.Logic/State/LoginState.cs
@@ -41,7 +41,6 @@ namespace PoGo.NecroBot.Logic.State
                 if (session.Settings.AuthType == AuthType.Google || session.Settings.AuthType == AuthType.Ptc)
                 {
                     await session.Client.Login.DoLogin();
-                    await LogLoginHistory(session, cancellationToken); 
                 }
                 else
                 {
@@ -55,7 +54,7 @@ namespace PoGo.NecroBot.Logic.State
             {
                 throw ae.Flatten().InnerException;
             }
-            catch (LoginFailedException ex)
+            catch (APIBadRequestException)
             {
                 session.EventDispatcher.Send(new ErrorEvent
                 {
@@ -63,10 +62,11 @@ namespace PoGo.NecroBot.Logic.State
                 });
                 
                 await Task.Delay(2000, cancellationToken);
-                throw ex;
+                throw new LoginFailedException();
             }
             catch (Exception ex) when (ex is PtcOfflineException || ex is AccessTokenExpiredException)
             {
+
                 session.EventDispatcher.Send(new ErrorEvent
                 {
                     Message = session.Translation.GetTranslation(TranslationString.PtcOffline)
@@ -118,6 +118,10 @@ namespace PoGo.NecroBot.Logic.State
                 });
                 await Task.Delay(2000, cancellationToken);
                 Environment.Exit(0);
+            }
+            catch(OperationCanceledException)
+            {
+                //just continue login if this happen, most case is bot switching...
             }
             catch (InvalidProtocolBufferException ex) when (ex.Message.Contains("SkipLastField"))
             {
@@ -172,13 +176,21 @@ namespace PoGo.NecroBot.Logic.State
                     System.Environment.Exit(1);
                 }
             }
-            catch(CaptchaException ex)
+            catch (OperationCanceledException)
+            {
+                //just continue login if this happen, most case is bot switching...
+            }
+            catch (CaptchaException ex)
             {
                 throw ex;
             }
             catch(ActiveSwitchByRuleException)
             {
                 //sometime the switch active happen same time with login by token expired. we need ignore it 
+            }
+            catch (APIBadRequestException)
+            {
+                throw new LoginFailedException();
             }
             
             session.LoggedTime = DateTime.Now;
@@ -187,24 +199,6 @@ namespace PoGo.NecroBot.Logic.State
                 return new BotSwitcherState(this.pokemonToCatch);
             }
             return new LoadSaveState();
-        }
-
-        private async Task LogLoginHistory(ISession session, CancellationToken cancellationToken)
-        {
-            string logFile = $"config\\login{DateTime.Now:dd-MM-yyyy}.log";
-            //if(!File.Exists(logFile) )
-            //{
-            //    File.CreateText(logFile);
-            //}
-
-            await Task.Run(() =>
-            {
-                try {
-                    string username = session.Settings.AuthType == AuthType.Ptc ? session.Settings.PtcUsername : session.Settings.GoogleUsername;
-                    File.AppendAllText(logFile, $"{DateTime.Now:dd-MM-yyyy hh:mm:ss}\t\t{username}\r\n");
-                }
-                catch(Exception) { }
-            });
         }
 
         private static async Task CheckLogin(ISession session, CancellationToken cancellationToken)